requirements.txt
azure-functions
azure-storage-file-datalake==12.0.0
azure-identity